Removed compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/535



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@2222 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/ucdtest.cpp
@@ -572,7 +572,7 @@ void UnicodeTest::TestCharLength()
 		1, 0xe000
 	};
 	
-	int16_t i;
+	int32_t i;
 	UBool multiple;
 	for(i=0; i<sizeof(codepoint)/sizeof(codepoint[0]); i=i+2){
 		UChar32 c=codepoint[i+1];
